Rename clear-messages file to match function
