+++ packages/wonder-blocks-announcer/src/index.ts
@@ -1,4 +1,4 @@
 import {announceMessage, type AnnounceMessageProps} from "./announce-message";
-import {clearMessages} from "./clear-message";
+import {clearMessages} from "./clear-messages";
 
 export {announceMessage, type AnnounceMessageProps, clearMessages};
